requirements.txt
google-ads==25.1.0
boto3==1.35.99
pandas==2.2.3
pyyaml==6.0.2
google-auth-oauthlib==1.2.1

# Deploy dependency — installed in the Airflow environment, not pip-installed here
# apache-airflow>=2.7.0
